requirements.txt
requests==2.10.0
pytest==2.9.0
behave==1.2.5
PyHamcrest
Faker
pyyaml==3.11